+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## [2-alpha17-SNAPSHOT]
+- ignoring missing source maps till we figure this problem out
+ 
+
 ## [2-alpha16] - 2023-04-06
 - simplified the way forms are sent to Clay (avoiding `tap>`)
 - escaping printed values for html